+++ src/modules/constructor/localConstructor.js
@@ -1,0 +1,6 @@
+export const cityEntry = [];
+
+export function CurrentCity(city, unit) {
+  this.city = city;
+  this.unit = unit;
+}
